requirements.txt
apache-beam==2.16.0
avro-python3==1.9.1
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
crcmod==1.7
dill==0.3.0
docopt==0.6.2
fastavro==0.21.24
fasteners==0.15
future==0.18.2
google-api-core==1.15.0
google-apitools==0.5.28
google-auth==1.10.0
google-cloud-bigquery==1.17.1
google-cloud-bigtable==1.0.0
google-cloud-core==1.1.0
google-cloud-datastore==1.7.4
google-cloud-pubsub==1.0.2
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.12.3
grpcio==1.26.0
hdfs==2.5.8
httplib2==0.12.0
idna==2.8
mock==2.0.0
monotonic==1.5
nose==1.3.7
nose-xunitmp==0.4.1
numpy==1.17.4
oauth2client==3.0.0
pandas==0.24.2
parameterized==0.6.3
pbr==5.4.4
protobuf==3.11.1
pyarrow==0.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pydot==1.4.1
PyHamcrest==1.9.0
pymongo==3.10.0
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==3.13
requests==2.22.0
rsa==4.0
six==1.13.0
tenacity==5.1.5
urllib3==1.25.7
websocket-client==0.56.0
